requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
plotly>=5.10.0
yfinance>=0.2.0
numpy>=1.20.0
statsmodels>=0.13.0
curl-cffi>=0.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0